vyatta-routing-instance: Add python3-vrfmanager
Some features (eg. vyatta-show-platform-dataplane.py) use the
existence of the vrfmanager module to determine if an image has
routing instance support, and therefore do not have a hard dependency
on python3-vrfmanager.

In fact there are no packages in DANOS which have a hard dependency
on python3-vrfmanager which results in the package not being
installed and thus breaking this detection mechanism.

Therefore explicitly install python3-vrfmanager in DANOS images.

This also addresses an issue with the DHCP client overlap-subnet
script, for which there is an un-declared hard dependency on
python3-vrfmanager.

This should be an interim solution - in future we should adjust the
vrf-manager package dependencies to allow installation of the
supporting libraries without automatically depending on full VRF
support. In the case where VRF support is not available, the
libraries should fail gracefully.

VRVDR-53314