Ensure sandbox GIDs are consistent with host
The /etc/group file generated during the creation of the user sandbox
chroot filesystem will likely contain groups which also exist in the
host /etc/group file (eg. ssh).

However because the set of packages installed in the chroot
environment, and/or the package setup order differs, the group IDs
allocated for those groups can differ between the host system and
the sandbox environment.

We currently require the group name <--> ID mapping to be consistent
between sandboxes and the host. To ensure this we manipulate the
sandbox chroot /etc/group and update the IDs of common groups so they
match the host /etc/group.

VRVDR-52842