0990-create-chroot-fs.chroot: remove perl-doc
The version of mmdebstrap in Buster has a dependency on perl-doc and
as there is (currently) no other dependency on perl-doc, sandbox
creation results in the not insignificant ~15M package remaining on
the final image.

Using autoremove does not remove the dependency, likely due to other
packages suggesting perl-doc. Therefore dpkg is instead used to
remove perl-doc, which helps keep the image size down.

A failure of dpkg is ignored to avoid breakage if other packages do
come along which require perl-doc in the image.

It should be possible to remove this on rebase to Bullseye, as the
version of mmdebstrap merely suggests perl-doc.

VRVDR-53942